+++ src/c975LConfigBundle.php
@@ -9,9 +9,12 @@
 
 namespace c975L\ConfigBundle;
 
-use c975L\ConfigBundle\DependencyInjection\Compiler\AlertProviderPass;
-use c975L\ConfigBundle\DependencyInjection\Compiler\MenuProviderPass;
-use c975L\ConfigBundle\DependencyInjection\Compiler\WhatsNewProviderPass;
+use c975L\ConfigBundle\DependencyInjection\Compiler\TaggedInterfacePass;
+use c975L\ConfigBundle\Management\AlertProviderInterface;
+use c975L\ConfigBundle\Management\LinkableRouteProviderInterface;
+use c975L\ConfigBundle\Management\MenuProviderInterface;
+use c975L\ConfigBundle\Management\ShortcutProviderInterface;
+use c975L\ConfigBundle\Management\WhatsNewProviderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
@@ -20,9 +23,11 @@ class c975LConfigBundle extends AbstractBundle
 {
     public function build(ContainerBuilder $container): void
     {
-        $container->addCompilerPass(new MenuProviderPass());
-        $container->addCompilerPass(new WhatsNewProviderPass());
-        $container->addCompilerPass(new AlertProviderPass());
+        $container->addCompilerPass(new TaggedInterfacePass(MenuProviderInterface::class, 'c975l.management_menu_provider'));
+        $container->addCompilerPass(new TaggedInterfacePass(WhatsNewProviderInterface::class, 'c975l.whatsnew_provider'));
+        $container->addCompilerPass(new TaggedInterfacePass(AlertProviderInterface::class, 'c975l.alert_provider'));
+        $container->addCompilerPass(new TaggedInterfacePass(ShortcutProviderInterface::class, 'c975l.shortcut_provider'));
+        $container->addCompilerPass(new TaggedInterfacePass(LinkableRouteProviderInterface::class, 'c975l.linkable_route_provider'));
     }
 
     public function loadExtension(array $config, ContainerConfigurator $containerConfigurator, ContainerBuilder $containerBuilder): void